+++ airbyte-integrations/connectors/source-azure-blob-storage/setup.py
@@ -6,7 +6,7 @@
 from setuptools import find_packages, setup
 
 MAIN_REQUIREMENTS = [
-    "airbyte-cdk>=0.51.17",
+    "airbyte-cdk>=0.52.5",
     "smart_open[azure]",
     "pytz",
     "fastavro==1.4.11",
